+++ extensions/amp-a4a/0.1/friendly-frame-util.js
@@ -21,7 +21,7 @@ import {
   installFriendlyIframeEmbed,
   setFriendlyIframeEmbedVisible,
 } from '../../../src/friendly-iframe-embed';
-import {installUrlReplacementsForEmbed} from '../../../src/service/url-replacements-impl';
+import {installUrlReplacementsForEmbed} from '../../../src/service/url-replacements-for-embed';
 import {setStyle} from '../../../src/style';
 
 /**
